Ahmed Ehab Final TIMER module Test

+++ Timer_Draft.h
@@ -11,11 +11,10 @@
 #include "TIMER_cnfg.h"
 
 
-
-
 ACK TIMER_init(void);
-ACK Time_Delay ( TIMER_t Timer_Select , float Required_Delay  );
+ACK Time_Delay ( TIMER_t Timer_Select , double Required_Delay  );
 ACK TIMER_Start ( TIMER_t Timer_Select );
+
 
 
 #endif /* TIMER_H_ */
